+++ Kufar3/Repositories/CategoryRepository.cs
@@ -8,17 +8,17 @@ namespace Kufar3.Repositories
 {
     public class CategoryRepository : BaseRepository
     {
-        public Category GetById(int? id)
+        public Category GetById(long? id)
         {
             return Context.Categories.FirstOrDefault(x => x.Id == id);
         }
 
-        public SubCategory GetSubCategoryById(int? id)
+        public SubCategory GetSubCategoryById(long? id)
         {
             return Context.SubCategories.FirstOrDefault(x => x.Id == id);
         }
 
-        public IQueryable<SubCategory> GetSubCategoriesByCategoryId(int? id)
+        public IQueryable<SubCategory> GetSubCategoriesByCategoryId(long? id)
         {
             return Context.SubCategories.Where(x => x.CategoryId == id);
         }
